# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MDM Platform API - Mobile Device Management para Android TV
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  admin: "admin",
  operator: "operator",
} as const;

export interface RegisterRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  tenantId?: string;
  tenantName?: string;
  role?: RegisterRequestRole;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  admin: "admin",
  operator: "operator",
} as const;

export interface UserProfile {
  id: string;
  email: string;
  role: UserProfileRole;
  tenantId: string;
  tenantName?: string;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface MeResponse {
  user: UserProfile;
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  admin: "admin",
  operator: "operator",
} as const;

export interface CreateUserRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  role?: CreateUserRequestRole;
}

export type DeviceStatus = (typeof DeviceStatus)[keyof typeof DeviceStatus];

export const DeviceStatus = {
  online: "online",
  offline: "offline",
  unknown: "unknown",
} as const;

export interface Device {
  id: string;
  name: string;
  ip: string;
  status: DeviceStatus;
  tenantId: string;
  lastSeen?: string | null;
  createdAt: string;
}

export interface CreateDeviceRequest {
  name: string;
  ip: string;
}

export type PingResponseStatus =
  (typeof PingResponseStatus)[keyof typeof PingResponseStatus];

export const PingResponseStatus = {
  online: "online",
  offline: "offline",
} as const;

export interface PingResponse {
  deviceId: string;
  status: PingResponseStatus;
  online: boolean;
}

export type CommandRequestAction =
  (typeof CommandRequestAction)[keyof typeof CommandRequestAction];

export const CommandRequestAction = {
  screen_toggle: "screen_toggle",
  screen_on: "screen_on",
  screen_off: "screen_off",
  home: "home",
  back: "back",
  reboot: "reboot",
  open_app: "open_app",
  install_apk: "install_apk",
  uninstall_app: "uninstall_app",
  list_apps: "list_apps",
  sync_apps: "sync_apps",
  keyevent: "keyevent",
  kiosk_enable: "kiosk_enable",
} as const;

export interface CommandRequest {
  action: CommandRequestAction;
  param?: string;
  keycode?: number;
}

export type CommandResponseStatus =
  (typeof CommandResponseStatus)[keyof typeof CommandResponseStatus];

export const CommandResponseStatus = {
  success: "success",
  error: "error",
} as const;

export interface CommandResponse {
  commandId: string;
  status: CommandResponseStatus;
  output?: string;
  error?: string;
}

export type CommandStatus = (typeof CommandStatus)[keyof typeof CommandStatus];

export const CommandStatus = {
  pending: "pending",
  running: "running",
  success: "success",
  error: "error",
} as const;

export interface Command {
  id: string;
  deviceId: string;
  command: string;
  status: CommandStatus;
  response?: string | null;
  createdAt: string;
  completedAt?: string | null;
}

export interface App {
  id: string;
  deviceId: string;
  packageName: string;
  appName?: string | null;
  createdAt: string;
}

export interface Log {
  id: string;
  deviceId: string;
  message: string;
  level: string;
  createdAt: string;
}

export interface ScheduledTask {
  id: string;
  tenantId: string;
  deviceId?: string | null;
  name: string;
  cronExpression: string;
  action: string;
  actionParam?: string | null;
  enabled: boolean;
  createdAt: string;
}

export interface CreateScheduledTaskRequest {
  name: string;
  deviceId?: string;
  cronExpression: string;
  action: string;
  actionParam?: string;
  enabled?: boolean;
}

export type GetDeviceLogsParams = {
  limit?: number;
};

export type GetAllLogsParams = {
  limit?: number;
};