# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MDM Platform API - Mobile Device Management para Android TV
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Iniciar sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Credenciales inválidas
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Registrar nuevo usuario y tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Usuario registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email ya registrado
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Obtener perfil del usuario autenticado
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Perfil del usuario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
  /devices:
    get:
      operationId: getDevices
      tags: [devices]
      summary: Listar dispositivos del tenant
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de dispositivos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
    post:
      operationId: createDevice
      tags: [devices]
      summary: Registrar nuevo dispositivo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeviceRequest"
      responses:
        "201":
          description: Dispositivo creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{id}:
    get:
      operationId: getDevice
      tags: [devices]
      summary: Obtener dispositivo por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dispositivo encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          description: No encontrado
    put:
      operationId: updateDevice
      tags: [devices]
      summary: Actualizar dispositivo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeviceRequest"
      responses:
        "200":
          description: Dispositivo actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
    delete:
      operationId: deleteDevice
      tags: [devices]
      summary: Eliminar dispositivo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Eliminado
...
  /devices/{id}/command:
    post:
      operationId: sendCommand
      tags: [commands]
      summary: Enviar comando ADB al dispositivo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommandRequest"
      responses:
        "200":
          description: Resultado del comando
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResponse"
  /devices/{id}/commands:
    get:
      operationId: getDeviceCommands
      tags: [commands]
      summary: Historial de comandos del dispositivo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lista de comandos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Command"
...
  /devices/{id}/logs:
    get:
      operationId: getDeviceLogs
      tags: [logs]
      summary: Obtener logs del dispositivo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Logs del dispositivo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Log"
  /logs:
    get:
      operationId: getAllLogs
      tags: [logs]
      summary: Obtener todos los logs del tenant
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Logs del tenant
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Log"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        tenantId:
          type: string
        tenantName:
          type: string
        role:
          type: string
          enum: [admin, operator]
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    MeResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - user
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, operator]
        tenantId:
          type: string
        tenantName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - role
        - tenantId
...
    Device:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        ip:
          type: string
        status:
          type: string
          enum: [online, offline, unknown]
        tenantId:
          type: string
        lastSeen:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - ip
        - status
        - tenantId
        - createdAt
    CreateDeviceRequest:
      type: object
      properties:
        name:
          type: string
        ip:
          type: string
      required:
        - name
        - ip
...
    CommandRequest:
      type: object
      properties:
        action:
          type: string
          enum:
            - screen_toggle
            - screen_on
            - screen_off
            - home
            - back
            - reboot
            - open_app
            - install_apk
            - uninstall_app
            - list_apps
            - sync_apps
            - keyevent
            - kiosk_enable
        param:
          type: string
        keycode:
          type: integer
      required:
        - action
    CommandResponse:
      type: object
      properties:
        commandId:
          type: string
        status:
          type: string
          enum: [success, error]
        output:
          type: string
        error:
          type: string
      required:
        - commandId
        - status
    Command:
      type: object
      properties:
        id:
          type: string
        deviceId:
          type: string
        command:
          type: string
        status:
          type: string
          enum: [pending, running, success, error]
        response:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - deviceId
        - command
        - status
        - createdAt
...
    Log:
      type: object
      properties:
        id:
          type: string
        deviceId:
          type: string
        message:
          type: string
        level:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - deviceId
        - message
        - level
        - createdAt
...